Remove Client inheritance
